requirements.txt
aiogram
asyncio
pip
python-dotenv